Harden inventory cache plugin implementation

+++ AsfInventoryCachePlugin/Models/InventorySnapshot.cs
@@ -3,7 +3,7 @@ namespace AsfInventoryCachePlugin.Models;
 public sealed class InventorySnapshot {
 	public string BotName { get; init; } = string.Empty;
 	public DateTimeOffset UpdatedAtUtc { get; init; }
-	public int TotalTradableAssets { get; init; }
+	public long TotalTradableAssets { get; init; }
 	public int UniqueAssetKeys { get; init; }
 	public List<InventorySnapshotEntry> Entries { get; init; } = [];
 }
@@ -12,6 +12,6 @@ public sealed class InventorySnapshotEntry {
 	public uint RealAppID { get; init; }
 	public string Type { get; init; } = string.Empty;
 	public ulong ClassID { get; init; }
-	public int Count { get; init; }
+	public long Count { get; init; }
 	public string? Name { get; init; }
 }
